requirements.txt
tableauserverclient==0.7
keyring==13.2.1
GitPython==2.1.11
